Add support for some control sequences

+++ src/spatial/point.rs
@@ -20,10 +20,10 @@ impl<Unit> Point<Unit> {
         }
     }
 
-    pub fn shifted(&self, horizontal_shift: u32, vertical_shift: u32) -> Point<Unit> {
-        let horizontal_distance = self.horizontal_distance + horizontal_shift;
-        let vertical_distance = self.vertical_distance + vertical_shift;
-        Point::new(horizontal_distance, vertical_distance)
+    pub fn shifted(&self, horizontal_shift: i32, vertical_shift: i32) -> Point<Unit> {
+        let horizontal_distance = self.horizontal_distance as i32 + horizontal_shift;
+        let vertical_distance = self.vertical_distance as i32 + vertical_shift;
+        Point::new(horizontal_distance as u32, vertical_distance as u32)
     }
 
     pub fn with_origin(&self, origin: Point<Unit>) -> Point<Unit> {
@@ -41,10 +41,10 @@ impl<Unit> Point<Unit> {
     }
 }
 
-impl<Unit> Add<(u32, u32)> for Point<Unit> {
+impl<Unit> Add<(i32, i32)> for Point<Unit> {
     type Output = Point<Unit>;
 
-    fn add(self, rhs: (u32, u32)) -> Self::Output {
+    fn add(self, rhs: (i32, i32)) -> Self::Output {
         self.shifted(rhs.0, rhs.1)
     }
 }
